Add static instance of api in sahil name sorter web
